+++ src/pyochain/collections/__init__.py
@@ -1,5 +1,6 @@
 """Additional collection types."""
 
+from ._deque import Deque
 from ._main import StableSet
 
-__all__ = ["StableSet"]
+__all__ = ["Deque", "StableSet"]